requirements.txt
django==1.11.17
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.4.6
mongoengine==0.9.0
pymongo==2.8.1
python-json-logger==0.1.2
pytz==2019.2
